requirements.txt
google-cloud-bigquery==3.27.0
pandas==2.2.3
openai==1.54.4
jsonlines==4.0.0
db-dtypes==1.3.1